Fixed critical typo, Raycasting improvements
Fixed a critical typo causing both the ray count and ray bounce count to be set to the config file's ray count value

Changed the raycast distribution from a golden spiral to the more effective offset Fibonacci lattice. (see http://extremelearning.com.au/how-to-evenly-distribute-points-on-a-sphere-more-effectively-than-the-canonical-fibonacci-lattice/) This should improve render performance by ensuring an even distribution of rays.